LayoutTests/http/tests/xmlhttprequest/access-control-sandboxed-iframe-denied-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-sandboxed-iframe-denied.cgi. Cannot make any requests from null.
This test verifies that sandboxed iframe does not have XmlHttpRequest access to its server. It will print "PASS" on success.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: Exception thrown. Sandboxed iframe XHR access was denied in 'send'. [NETWORK_ERR: XMLHttpRequest Exception 101].
